Do not create create a new `ConnectionFactory` if one already exists in parent
Fix #63

Make sure that the creation of a new `ConnectionFactory` when
cloud connectors are used is done only when the binder
context is not a child of the main context, otherwise inherit
the connection factory as in normal case. This ensures that its lifecycle
is in sync with the parent.