SCDF-GH-1528: Fix DLQ Routing Key for SCDF
Fixes spring-cloud/spring-cloud-dataflow#1528

Data flow creates destinations with name `<stream>.module` and sets the group to the stream name.
The rabbit binder binders a queue `<destination>.<group>` to the `<destination>` exchange.
If `autoBindDlq` is set, a DLQ is created and bound with key matching the original queue name.

There is logic manipulating the destination name to strip off the prefix (if any) because it's re-added later.

This logic doesn't work with SCDF because the group occurs twice in the queue name; we end up routing on
just the `<group>`.

Remove the strip and add back of the prefix to avoid this logic altogether.

Modify one of the tests to simulate SCDF naming.